Disallow pushing
It's a bit annoying when trying to vine or something and you both get pushed off because of it.

I just tested the tag hitbox with this and it's relatively the same, though it is slightly easier to tag due to not being able to push them. Distance is the same though so I don't see any issues with it.