Restructure mutes to use ban structure
The old structure for mutes was a bit... all over the shop, admittedly. The ban structure is very easy to understand and consistency is key after all.

Formerly, mutes were stored in the server and would reset with the server respectively which caused players to be unmuted.

Note: Server won't start unless you make a mutes.txt file, after that there are no issues.