Fix flag dropping when player walks over spawn
This bug has been annoying me for a while now. I've known about it for a long time and have tested many hypotheses but none of which proved correct until now.

The issue was set in place for the classic days when spectators had their own spawns. Now that this is redundant, this code isn't really needed anymore.