Fix flag not respawning when dropped above ceiling
It looks as if the player was disconnected when placing above the build ceiling. The flag also didn't respawn when killed during stalemate mode. I've made it so the server automatically places the flag at the build ceiling if the player is above it. Functionality for regular dropping is still the same.