Disallow store before game start in tourney mode
I've seen some folks (and myself) accidentally press a store hotkey before the game starts and it wastes points so I just prevented the use of store altogether until the game starts.

There is a bit of inconsistent formatting as the same message when placing a block is formatted in green, though the other messages are formatted in yellow. For consistency with the other prerequisites, I've just used yellow.