Flame should start 1 block farther when ping > 90
The 90ms value may need to be tweaked, but the intention behind this is to make flamethrower as balanced as possible. Since the tick speed is set so high, players with higher ping are at a serious disadvantage when it comes to flamethrower. Often, players will swim inside of flames and you can even get stuck in your own flamethrower.