Add isCreepering flag to Player class
This flag is reset by default and is reset when:
* the player dies
* the player joins any team
* the game ends
* a new game starts
* the player blows himself up due to creeper

This flag is set after the creeper item is successfully bought.
The CreeperTask will check if the isCreepering flag is true and if the
player's session is still registered with the world before actually
doing the explosion.

Also, in the unusual case where the creeper delay is longer than the
creeper cooldown, the creeper item cannot be bought after the cooldown
expires but before the delay ends.