Enforce TNT lag pillaring (not TNT rushing)
The difference between the two is that lag pillaring abuses the block revert when you already have a TNT. TNT rushing is just exploding the TNT below you with the detonator but due to lag, it /looks/ like it's the same thing.

I've implemented a lag TNT counter to auto-kick players if they continue to lag pillar. If there's an accidental lag pillar during a TNT rush, the player will be notified.